DiAna: DIsproportionality ANAlysis and Pharmacovigilance in the FAERS

Advanced tools for analyzing the FDA Adverse Event Reporting System (FAERS) data, enhancing drug safety research. Provides functions to download cleaned FAERS data and import relational datasets, facilitating case identification and control selection to address confounding in analyses. Provides functions for descriptive and disproportionality analysis, time-to-onset analysis, network analysis, retrieval algorithms, and visualization.

See also

Author

Maintainer: Michele Fusaroli michele.fusaroli@gmail.com (ORCID) [copyright holder]

Authors:

Valentina Giunchi valentina.giunchi2@unibo.it (ORCID)

Other contributors:

Violetta Zoffoli [contributor]

Lionel Van Holle lionel.vanholle@opensourcepv.com (ORCID) [contributor]

Takamasa Sakai tksakai@meijo-u.ac.jp [contributor]

Nhung Trinh nhung.trinh@farmasi.uio.no [contributor]