Generate Descriptive Statistics for a Sample

This function generates descriptive statistics for a sample of data, including cases and non-cases, and saves the results to an Excel file.

Usage

descriptive(
  pids_cases,
  RG = NULL,
  drug = NULL,
  save_in_excel = TRUE,
  file_name = "Descriptives.xlsx",
  vars = c("sex", "Submission", "Reporter", "age_range", "Outcome", "country",
    "continent", "age_in_years", "wt_in_kgs", "Reactions", "Indications", "Substances",
    "year", "role_cod", "time_to_onset"),
  list_pids = list(),
  method = "independence_test",
  temp_demo = Demo,
  temp_drug = Drug,
  temp_reac = Reac,
  temp_indi = Indi,
  temp_outc = Outc,
  temp_ther = Ther
)

Arguments

pids_cases

A vector of primary IDs for cases.

RG

A vector of primary IDs: reference group. Default is NULL.

drug

A vector of drug names. Default is NULL.

save_in_excel

Whether to save the outcome in an excel. Defaults to TRUE

file_name

The name of the Excel file to save the results. Default is "Descriptives.xlsx". It only works if save_in_excel is TRUE.

vars

A character vector of variable names to include in the analysis.

list_pids

A list of vectors with primary IDs for custom groups whose distribution should be described. Default is an empty list.

method

The method for Chi-square test analysis, either "independence_test" or "goodness_of_fit". Default is "independence_test". It applies only for comparisons between cases and non-cases.

temp_demo

Demo dataset. Defaults to Demo. Can be se to sample_Demo for testing

temp_drug

Drug dataset. Can be set to sample_Drug for testing

temp_reac

Reac dataset. Can be set to sample_Reac for testing

temp_indi

Indi dataset. Can be set to sample_Indi for testing

temp_outc

Outc dataset. Can be set to sample_Outc for testing

temp_ther

Ther dataset. Can be set to sample_Ther for testing

Value

The function generates descriptive statistics as a gt_table and potentially saves them to an Excel file.

Examples

if (FALSE) { # \dontrun{
import("DEMO", quarter = "23Q1")
pids_cases <- sample(Demo$primaryid, 100)
RG <- sample(Demo[!primaryid %in% pids_cases]$primaryid, 100)

# Generate descriptive statistics for cases
descriptive(pids_cases = pids_cases)
descriptive(pids_cases = pids_cases, drug = "paracetamol")

# Generate descriptive statistics for cases and non-cases
descriptive(pids_cases = pids_cases, RG = RG)
descriptive(pids_cases = pids_cases, RG = RG, drug = "paracetamol")
} # }