Reporting rates of events or substances

counts the occurrences of reactions/indications/substance for a given set of primary IDs. Calculates % as the proportion of individuals recording the event.

Usage

reporting_rates(
  pids_cases,
  entity = "reaction",
  level = "pt",
  drug_role = c("PS", "SS", "I", "C"),
  drug_indi = NA,
  temp_reac = Reac,
  temp_drug = Drug,
  temp_indi = Indi
)

Arguments

pids_cases

Vector of primary IDs to consider for counting reactions.

entity

Entity investigated. It can be one of the following:

reaction;

indication;

substance.

level

The desired MedDRA or ATC level for counting (default is "pt").

drug_role

is only used for substances. By default both suspect and concomitant drugs are included.

drug_indi

is only used for indications. By default the indications of all the drugs of the selected primaryids are considered, but you can specify a vector of drugs.

temp_reac

Reac dataset. Can be set to sample_Reac for testing

temp_drug

Drug dataset. Can be set to sample_Drug for testing

temp_indi

Indi dataset. Can be set to sample_Indi for testing

Value

A data.table containing counts and percentages of the investigated entity at the specified level and in descending order.

Examples

if (FALSE) {
# select only reports recording paracetamol from the sample dataset,
# and provide the most reported events (at the hlt level),
# drugs (at the ATC class3 level), and indications (at the PT level).
pids_cases <- unique(sample_Drug[substance == "paracetamol"]$primaryid)
reporting_rates(pids, "reaction", "hlt",
  temp_reac = sample_Reac,
  temp_drug = sample_Drug, temp_indi = sample_Indi
)
reporting_rates(pids, "indication", "pt",
  temp_reac = sample_Reac,
  temp_drug = sample_Drug, temp_indi = sample_Indi
)
reporting_rates(pids,
  entity = "substance", level = "Class3",
  temp_reac = sample_Reac, temp_drug = sample_Drug,
  temp_indi = sample_Indi
)
}